Create about-added-for-mysql-trigger.php
After toying with the syntax and different suggestions from many other stackoverflow questions/answers, I finally found the right combination of small code additions to have automatic "created" and "updated" columns in my database table.

In my hook_schema() function:code below;
In my hook_install() function: code below(2);
The 'triggername' is whatever you want to name your trigger. The 'tablename' is the name of the table you established in hook_schema() that has the created and updated columns.

Hopefully this can be of use to somebody. When I was looking into this a couple weeks ago, there were no posts with someone doing this within drupal and it was very difficult to figure out how to adapt the different suggestions to the drupal framework. This is the best solution I figured out. If you have another one you believe to be better go ahead and post it!